Stub-out Python API for adding subcomponents
