+++ src/sst/core/model/Makefile.inc
@@ -8,7 +8,11 @@ sst_core_sources += \
 	model/sstmodel.h \
 	model/sstmodel.cc \
 	model/pymodel.h \
-	model/pymodel.cc
+	model/pymodel.cc \
+	model/pymodel_link.h \
+	model/pymodel_link.cc \
+	model/pymodel_comp.h \
+	model/pymodel_comp.cc
 
 libexec_SCRIPTS = model/xmlToPython.py
 EXTRA_DIST += model/xmlToPython.py
